+++ v3/internal/integrationsupport/integrationsupport.go
@@ -34,6 +34,7 @@ func AddAgentSpanAttribute(txn *newrelic.Transaction, key string, val string) {
 const (
 	testLicenseKey = "0123456789012345678901234567890123456789"
 	SampleAppName  = "my app"
+	TestEntityGUID = "testEntityGUID123"
 )
 
 // ExpectApp combines Application and Expect, for use in validating data in test apps
@@ -104,4 +105,6 @@ var AppLogEnabledCfgFn = func(cfg *newrelic.Config) {
 // SampleEverythingReplyFn is a reusable ConnectReply function that samples everything
 var SampleEverythingReplyFn = func(reply *internal.ConnectReply) {
 	reply.SetSampleEverything()
+	reply.EntityGUID = TestEntityGUID
+	reply.EventData = internal.DefaultEventHarvestConfig(internal.MaxTxnEvents, internal.MaxLogEvents, internal.MaxCustomEvents)
 }